Refactor Acas
