+++ src/contractTest/java/uk/gov/hmcts/reform/et/syaapi/consumer/ccd/StartForCitizenConsumerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.et.syaapi.consumer.test.ccd;
+package uk.gov.hmcts.reform.et.syaapi.consumer.ccd;
 
 import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
 import au.com.dius.pact.consumer.junit5.PactTestFor;
@@ -9,14 +9,14 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
-import uk.gov.hmcts.reform.et.syaapi.consumer.test.SpringBootContractBaseTest;
+import uk.gov.hmcts.reform.et.syaapi.consumer.SpringBootContractBaseTest;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.springframework.http.HttpMethod.GET;
 
-class StartTestForCitizenConsumer extends SpringBootContractBaseTest {
+class StartForCitizenConsumerTest extends SpringBootContractBaseTest {
 
     @Pact(provider = "ccd_data_store_api_cases", consumer = "et_sya_api_service")
         RequestResponsePact startForCitizen(PactDslWithProvider builder) {
