+++ src/contractTest/java/uk/gov/hmcts/reform/et/syaapi/consumer/idam/S2SConsumerTest.java
@@ -1,7 +1,7 @@
-package uk.gov.hmcts.reform.et.syaapi.consumer.test.idam;
+package uk.gov.hmcts.reform.et.syaapi.consumer.idam;
 
 import au.com.dius.pact.consumer.MockServer;
-import au.com.dius.pact.consumer.dsl.DslPart;
+import au.com.dius.pact.consumer.dsl.PactDslJsonBody;
 import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
 import au.com.dius.pact.consumer.junit5.PactConsumerTestExt;
 import au.com.dius.pact.consumer.junit5.PactTestFor;
@@ -9,7 +9,6 @@ import au.com.dius.pact.core.model.RequestResponsePact;
 import au.com.dius.pact.core.model.annotations.Pact;
 import au.com.dius.pact.core.model.annotations.PactDirectory;
 import io.restassured.RestAssured;
-import org.assertj.core.api.Assertions;
 import org.json.JSONObject;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -19,24 +18,25 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import java.util.Map;
 
-import static io.pactfoundation.consumer.dsl.LambdaDsl.newJsonBody;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.http.HttpMethod.GET;
 import static org.springframework.http.HttpStatus.OK;
 
 @ExtendWith(PactConsumerTestExt.class)
 @ExtendWith(SpringExtension.class)
 @PactDirectory("pacts")
-class IdamConsumerTest {
-    private static final String IDAM_USER_DETAILS = "/details";
+class S2SConsumerTest {
+    private static final String S2S_URL = "/lease";
 
-    @Pact(provider = "idam_user_details", consumer = "et_sya_api_service")
-    RequestResponsePact executeIdamUserDetailApi(PactDslWithProvider builder) {
-        Map<String, String> responseHeaders = Map.of(HttpHeaders.AUTHORIZATION, "Bearer UserAuthToken");
+    @Pact(provider = "s2-auth-api", consumer = "et_sya_api_service")
+    RequestResponsePact executeServiceAuthApiGetToken(PactDslWithProvider builder) {
+
+        Map<String, String> responseHeaders = Map.of(HttpHeaders.AUTHORIZATION, "someToken");
 
         return builder
-            .given("a user exists")
-            .uponReceiving("Provider receives a token request and send user details to the API")
-            .path(IDAM_USER_DETAILS)
+            .given("a secret exists")
+            .uponReceiving("Provider receives a token request request from et-sya-api API")
+            .path(S2S_URL)
             .method(GET.toString())
             .willRespondWith()
             .status(OK.value())
@@ -46,15 +46,16 @@ class IdamConsumerTest {
     }
 
     @Test
-    @PactTestFor(pactMethod = "executeIdamUserDetailApi")
-    void shouldReceiveUserDetails(MockServer mockServer) {
+    @PactTestFor(pactMethod = "executeServiceAuthApiGetToken")
+    void shouldReceiveTokenAnd200(MockServer mockServer) {
 
         String responseBody = RestAssured
             .given()
             .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+            .body(createRequestBody())
             .log().all(true)
             .when()
-            .get(mockServer.getUrl() + IDAM_USER_DETAILS)
+            .get(mockServer.getUrl() + S2S_URL)
             .then()
             .statusCode(200)
             .and()
@@ -62,17 +63,20 @@ class IdamConsumerTest {
             .asString();
 
         JSONObject response = new JSONObject(responseBody);
-        Assertions.assertThat(response).isNotNull();
+        assertThat(response).isNotNull();
+        assertThat(response.getString("token")).isNotBlank();
     }
 
-    private DslPart createAuthResponse() {
-        return newJsonBody(o -> o.stringType("id",
-                                         "123432")
-            .stringType("forename", "Joe")
-            .stringType("surname", "Bloggs")
-            .stringType("email", "joe.bloggs@hmcts.net")
-            .booleanType("active", true)
-            .array("roles", role -> role.stringType("caseworker").stringType("citizen"))).build();
+
+    private PactDslJsonBody createAuthResponse() {
+        return new PactDslJsonBody()
+            .stringType("token","someMicroServiceToken");
     }
+
+    private static String createRequestBody() {
+        return new StringBuffer().append("{\"microservice\": \"microServiceName\"))")
+            .append(" \"oneTimePassword\": \"987651\",")
+            .append(" }").toString();
+    }
+
 }
-
